delete할 때 checkCurrentUserIsTheAuthor 체크

+++ back/realWorld/src/main/java/study/realWorld/service/ArticlesServiceImpl.java
@@ -43,6 +43,7 @@ public class ArticlesServiceImpl implements ArticlesService {
     @Transactional
     public void deleteBySlug(String slug) {
         Articles articles = getArticleBySlugOr404(slug);
+        checkCurrentUserIsTheAuthor(articles);
         articlesRepository.delete(articles);
     }
